requirements.txt
pandas
fastapi
dotenv
pydantic
uvicorn
pymysql
sqlalchemy
bcrypt
pyjwt
alembic